refactor: 댓글, 대댓글 validation
- 공백 입력 안되게 변경
