+++ ToyProject3/src/main/java/team7/example/ToyProject3/dto/reply/ReplyRequestDto.java
@@ -4,10 +4,9 @@ import javax.validation.constraints.NotBlank;
 
 import lombok.Builder;
 import lombok.Getter;
-import team7.example.ToyProject3.domain.Board;
 import team7.example.ToyProject3.domain.Reply;
-import team7.example.ToyProject3.domain.User;
-
+import team7.example.ToyProject3.domain.board.Board;
+import team7.example.ToyProject3.domain.user.User;
 
 public class ReplyRequestDto {
