+++ ToyProject3/src/main/java/team7/example/ToyProject3/repository/ReplyRepository.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import team7.example.ToyProject3.domain.Board;
 import team7.example.ToyProject3.domain.Reply;
 
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
