+++ ToyProject3/src/main/java/team7/example/ToyProject3/service/ReplyService.java
@@ -1,8 +1,9 @@
 package team7.example.ToyProject3.service;
 
 import java.util.List;
+
+import team7.example.ToyProject3.domain.user.User;
 import team7.example.ToyProject3.domain.Reply;
-import team7.example.ToyProject3.domain.User;
 import team7.example.ToyProject3.dto.reply.ReplyRequestDto;
 import team7.example.ToyProject3.dto.reply.ReplyResponseDto;
